daemon: Add --enable-bpf-masquerade flag
Currently, the flag is noop except that it disables provisioning of
iptables MASQUERADE rules when set.

The flag is going to be used to control BPF MASQUERADE'ing.

By default, it is set to false, so that users who will upgrade to
Cilium v1.8 will keep the existing behaviour (= MASQ is done by
iptables).

Signed-off-by: Martynas Pumputis <m@lambda.lt>