vagrant: Do not require env. varables for all Vagrant cmds
Currently, to e.g. destroy a Vagrant VM, it is necessary to set some of
the same options that were used to create it (e.g., K8S, NWORKERS, and
now with the previous commit, NETNEXT). This commit fixes this by always
defining common VM names, so that commands like 'vagrant destroy' or
'vagrant ssh' work without setting env. variables.

Signed-off-by: Paul Chaignon <paul@cilium.io>