fragmap, policymap: clean-up map size configuration
As a follow-up to IPv4 fragmentation cleanup (#10927), improve the way
we handle configuration for the maximum number of entries in policy map
and fragment-tracking map.

Functional change: When the option passed by the user is outside of the
limits used for policy map (upper/lower) or fragmap (upper), fall back
to the limit and print a warning instead of stopping with an error.

Move such "limits" to package "defaults" (We cannot move them to the
respective map packages, as those cannot be imported from "config").

Remove unnecessary InitMapInfo() for fragmap. Keep it for policy map,
because passing the size at map creation and removing this function
would mean:

- Updating the list of arguments passed to Open(), OpenAndCreate(),
  Create() and pass the size (option.Config.PolicyMapMaxEntries) each
  time, which seems verbose and not really useful. We cannot import the
  "option" package in "policy" to use the option just once in newMap().

- Making sure that policymap.MaxEntries is never called from other
  packages before the map is created. This is probably not the case so
  we do want to set the size before map creation.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>